Move Sequential and slice_model to common.py
